+++ winui3/SquirrelNotifier.WinUI3/Helpers/ReviewEventParser.cs
@@ -8,7 +8,7 @@ using System.Text.Json;
 using SquirrelNotifier.WinUI3.Helpers;
 using SquirrelNotifier.WinUI3.Models;
 
-namespace SquirrelNotifier.WinUI3.Services;
+namespace SquirrelNotifier.WinUI3.Helpers;
 
 internal static class ReviewEventParser
 {