+++ winui3/SquirrelNotifier.WinUI3/Services/AutoPauseGate.cs
@@ -3,6 +3,7 @@
 // </copyright>
 
 using System.Globalization;
+using SquirrelNotifier.WinUI3.Helpers;
 using SquirrelNotifier.WinUI3.Models;
 
 namespace SquirrelNotifier.WinUI3.Services;